Rework map
