+++ src/constants/images.tsx
@@ -1,7 +1,13 @@
 import china from "../assets/china.jpg";
 import bubbletea from "../assets/bubble-tea.jpg";
+import farm from "../assets/farm.jpg";
+import cave from "../assets/cave.png";
+import cyrstalCave from "../assets/crystalCave.jpg";
 
 export default {
   bubbletea,
   china,
+  farm,
+  cave,
+  cyrstalCave,
 };
