+++ src/container/Map/Map.tsx
@@ -1,28 +1,14 @@
 import React from "react";
 import "./Map.css";
+import images from "../../constants/images";
+import MapAreaComponent from "../../components/MapArea/MapAreaComponent";
 
 export const Map = () => {
-  const handleOnClick = (e: { preventDefault: () => void }) => {
-    console.log("You have clicked in the specified area");
-  };
-
   return (
-    <div>
-      <img
-        width="600px"
-        src="https://www.nationsonline.org/maps/china_provinces_map1200.jpg"
-        alt="Map"
-        useMap="#workmap"
-      />
-      <map id="workmap" name="workmap">
-        <area
-          shape="rect"
-          coords="200,200,300,300"
-          alt="test"
-          href="https://c1.staticflickr.com/5/4052/4503898393_303cfbc9fd_b.jpg"
-          onClick={handleOnClick}
-        ></area>
-      </map>
+    <div className="map__main">
+      <MapAreaComponent image={images.farm} text="Farm" information="Level 1" />
+      <MapAreaComponent image={images.cave} text="Cave" information="Level 5" />
+      <MapAreaComponent image={images.cyrstalCave} text="CrystalCave" information="Level 10" />
     </div>
   );
 };
